Refactor para package data para ficar mais organizado
